+++ src/braket/default_simulator/density_matrix_simulator.py
@@ -179,7 +179,7 @@ class DensityMatrixSimulator(BaseLocalSimulator):
                         ],
                         "supportPhysicalQubits": False,
                         "supportsPartialVerbatimBox": False,
-                        "requiresContiguousQubitIndices": True,
+                        "requiresContiguousQubitIndices": False,
                         "requiresAllQubitsMeasurement": False,
                         "supportsUnassignedMeasurements": True,
                         "disabledQubitRewiringSupported": False,