+++ src/braket/default_simulator/simulator.py
@@ -21,6 +21,7 @@ import numpy as np
 from braket.device_schema import DeviceActionType
 from braket.ir.jaqcd import Program as JaqcdProgram
 from braket.ir.jaqcd.program_v1 import Results
+from braket.ir.jaqcd.shared_models import MultiTarget, OptionalMultiTarget
 from braket.ir.openqasm import Program as OpenQASMProgram
 from braket.task_result import (
     AdditionalMetadata,
@@ -128,7 +129,6 @@ class BaseLocalSimulator(OpenQASMSimulator):
 
         Args:
             circuit_ir (Union[OpenQASMProgram, JaqcdProgram]): Circuit specification.
-            qubit_count (int, jaqcd-only): Number of qubits.
             shots (int, optional): The number of shots to simulate. Default is 0, which
                 performs a full analytical simulation.
             batch_size (int, optional): The size of the circuit partitions to contract,
@@ -255,6 +255,7 @@ class BaseLocalSimulator(OpenQASMSimulator):
         openqasm_ir: OpenQASMProgram,
         simulation: Simulation,
         measured_qubits: list[int] = None,
+        mapped_measured_qubits: list[int] = None,
     ) -> GateModelTaskResult:
         return GateModelTaskResult.construct(
             taskMetadata=TaskMetadata(
@@ -266,20 +267,13 @@ class BaseLocalSimulator(OpenQASMSimulator):
                 action=openqasm_ir,
             ),
             resultTypes=results,
-            measurements=self._formatted_measurements(simulation, measured_qubits),
-            measuredQubits=(
-                measured_qubits if measured_qubits else self._get_all_qubits(simulation.qubit_count)
-            ),
+            measurements=self._formatted_measurements(simulation, mapped_measured_qubits),
+            measuredQubits=(measured_qubits or list(range(simulation.qubit_count))),
         )
 
     @staticmethod
-    def _validate_operation_qubits(operations: list[Operation]) -> None:
-        qubits_referenced = {target for operation in operations for target in operation.targets}
-        if qubits_referenced and max(qubits_referenced) >= len(qubits_referenced):
-            raise ValueError(
-                "Non-contiguous qubit indices supplied; "
-                "qubit indices in a circuit must be contiguous."
-            )
+    def _get_qubits_referenced(operations: list[Operation]) -> set[int]:
+        return {target for operation in operations for target in operation.targets}
 
     @staticmethod
     def _validate_result_types_qubits_exist(
@@ -353,10 +347,6 @@ class BaseLocalSimulator(OpenQASMSimulator):
                         raise NameError(f"Missing input variable '{missing_input}'.")
 
     @staticmethod
-    def _get_all_qubits(qubit_count: int) -> list[int]:
-        return list(range(qubit_count))
-
-    @staticmethod
     def _tensor_product_index_dict(
         observable: TensorProduct, func: Callable[[Observable], Any]
     ) -> dict[int, Any]:
@@ -387,6 +377,150 @@ class BaseLocalSimulator(OpenQASMSimulator):
             return str(observable.__class__.__name__)
 
     @staticmethod
+    def _map_circuit_to_contiguous_qubits(circuit: Union[Circuit, JaqcdProgram]) -> dict[int, int]:
+        """
+        Maps the qubits in operations and result types to contiguous qubits.
+
+        Args:
+            circuit (Union[Circuit, JaqcdProgram]): The circuit containing the operations and
+            result types.
+
+        Returns:
+            dict[int, int]: Map of qubit index to corresponding contiguous index
+        """
+        circuit_qubit_set = BaseLocalSimulator._get_circuit_qubit_set(circuit)
+        qubit_map = BaseLocalSimulator._contiguous_qubit_mapping(circuit_qubit_set)
+        BaseLocalSimulator._map_circuit_qubits(circuit, qubit_map)
+        return qubit_map
+
+    @staticmethod
+    def _get_circuit_qubit_set(circuit: Union[Circuit, JaqcdProgram]) -> set[int]:
+        """
+        Returns the set of qubits used in the given circuit.
+
+        Args:
+            circuit (Union[Circuit, JaqcdProgram]): The circuit from which to extract the qubit set.
+
+        Returns:
+            set[int]: The set of qubits used in the circuit.
+        """
+        if isinstance(circuit, Circuit):
+            return circuit.qubit_set
+        else:
+            operations = [
+                from_braket_instruction(instruction) for instruction in circuit.instructions
+            ]
+            if circuit.basis_rotation_instructions:
+                operations.extend(
+                    from_braket_instruction(instruction)
+                    for instruction in circuit.basis_rotation_instructions
+                )
+            return BaseLocalSimulator._get_qubits_referenced(operations)
+
+    @staticmethod
+    def _map_circuit_qubits(circuit: Union[Circuit, JaqcdProgram], qubit_map: dict[int, int]):
+        """
+        Maps the qubits in operations and result types to contiguous qubits.
+
+        Args:
+            circuit (Circuit): The circuit containing the operations and result types.
+            qubit_map (dict[int, int]): The mapping from qubits to their contiguous indices.
+
+        Returns:
+            Circuit: The circuit with qubits in operations and result types mapped
+            to contiguous qubits.
+        """
+        if isinstance(circuit, Circuit):
+            BaseLocalSimulator._map_circuit_instructions(circuit, qubit_map)
+            BaseLocalSimulator._map_circuit_results(circuit, qubit_map)
+        else:
+            BaseLocalSimulator._map_jaqcd_instructions(circuit, qubit_map)
+        return circuit
+
+    @staticmethod
+    def _map_circuit_instructions(circuit: Circuit, qubit_map: dict):
+        """
+        Maps the targets of each instruction in the circuit to the corresponding qubits in the
+        qubit_map.
+
+        Args:
+            circuit (Circuit): The circuit containing the instructions.
+            qubit_map (dict): A dictionary mapping original qubits to new qubits.
+        """
+        for ins in circuit.instructions:
+            ins._targets = tuple([qubit_map[q] for q in ins.targets])
+
+    @staticmethod
+    def _map_circuit_results(circuit: Circuit, qubit_map: dict):
+        """
+        Maps the targets of each result in the circuit to the corresponding qubits in the qubit_map.
+
+        Args:
+            circuit (Circuit): The circuit containing the results.
+            qubit_map (dict): A dictionary mapping original qubits to new qubits.
+        """
+        for result in circuit.results:
+            if isinstance(result, (MultiTarget, OptionalMultiTarget)) and result.targets:
+                result.targets = [qubit_map[q] for q in result.targets]
+
+    @staticmethod
+    def _map_jaqcd_instructions(circuit: JaqcdProgram, qubit_map: dict):
+        """
+        Maps the attributes of each instruction in the JaqcdProgram to the corresponding qubits in
+        the qubit_map.
+
+        Args:
+            circuit (JaqcdProgram): The JaqcdProgram containing the instructions.
+            qubit_map (dict): A dictionary mapping original qubits to new qubits.
+        """
+        for ins in circuit.instructions:
+            BaseLocalSimulator._map_instruction_attributes(ins, qubit_map)
+
+        if hasattr(circuit, "results") and circuit.results:
+            for ins in circuit.results:
+                BaseLocalSimulator._map_instruction_attributes(ins, qubit_map)
+
+        if circuit.basis_rotation_instructions:
+            for ins in circuit.basis_rotation_instructions:
+                ins.target = qubit_map[ins.target]
+
+    @staticmethod
+    def _map_instruction_attributes(instruction, qubit_map: dict):
+        """
+        Maps the qubit attributes of an instruction from JaqcdProgram to the corresponding
+        qubits in the qubit_map.
+
+        Args:
+            instruction: The Jaqcd instruction whose qubit attributes need to be mapped.
+            qubit_map (dict): A dictionary mapping original qubits to new qubits.
+        """
+        if hasattr(instruction, "control"):
+            instruction.control = qubit_map.get(instruction.control, instruction.control)
+
+        if hasattr(instruction, "controls") and instruction.controls:
+            instruction.controls = [qubit_map.get(q, q) for q in instruction.controls]
+
+        if hasattr(instruction, "target"):
+            instruction.target = qubit_map.get(instruction.target, instruction.target)
+
+        if hasattr(instruction, "targets") and instruction.targets:
+            instruction.targets = [qubit_map.get(q, q) for q in instruction.targets]
+
+    @staticmethod
+    def _contiguous_qubit_mapping(qubit_set: set[int]) -> dict[int, int]:
+        """
+        Maping of qubits to contiguous integers. The qubit mapping may be discontiguous or
+        contiguous.
+
+        Args:
+            qubit_set (set[int]): List of qubits to be mapped.
+
+        Returns:
+            dict[int, int]: Dictionary where keys are qubits and values are contiguous integers.
+        """
+        return {q: i for i, q in enumerate(sorted(qubit_set))}
+
+    @staticmethod
     def _formatted_measurements(
         simulation: Simulation, measured_qubits: Union[list[int], None] = None
     ) -> list[list[str]]:
@@ -407,21 +541,16 @@ class BaseLocalSimulator(OpenQASMSimulator):
         ]
         #  Gets the subset of measurements from the full measurements
         if measured_qubits is not None and measured_qubits != []:
-            if any(qubit in range(simulation.qubit_count) for qubit in measured_qubits):
-                measured_qubits = np.array(measured_qubits)
-                in_circuit_mask = measured_qubits < simulation.qubit_count
-                measured_qubits_in_circuit = measured_qubits[in_circuit_mask]
-                measured_qubits_not_in_circuit = measured_qubits[~in_circuit_mask]
+            measured_qubits = np.array(measured_qubits)
+            in_circuit_mask = measured_qubits < simulation.qubit_count
+            measured_qubits_in_circuit = measured_qubits[in_circuit_mask]
+            measured_qubits_not_in_circuit = measured_qubits[~in_circuit_mask]
 
-                measurements_array = np.array(measurements)
-                selected_measurements = measurements_array[:, measured_qubits_in_circuit]
-                measurements = np.pad(
-                    selected_measurements, ((0, 0), (0, len(measured_qubits_not_in_circuit)))
-                ).tolist()
-            else:
-                measurements = np.zeros(
-                    (simulation.shots, len(measured_qubits)), dtype=int
-                ).tolist()
+            measurements_array = np.array(measurements)
+            selected_measurements = measurements_array[:, measured_qubits_in_circuit]
+            measurements = np.pad(
+                selected_measurements, ((0, 0), (0, len(measured_qubits_not_in_circuit)))
+            ).tolist()
         return measurements
 
     def run_openqasm(
@@ -451,8 +580,12 @@ class BaseLocalSimulator(OpenQASMSimulator):
                 are requested when shots>0.
         """
         circuit = self.parse_program(openqasm_ir).circuit
+        qubit_map = BaseLocalSimulator._map_circuit_to_contiguous_qubits(circuit)
         qubit_count = circuit.num_qubits
         measured_qubits = circuit.measured_qubits
+        mapped_measured_qubits = (
+            [qubit_map[q] for q in measured_qubits] if measured_qubits else None
+        )
 
         self._validate_ir_results_compatibility(
             circuit.results,
@@ -465,14 +598,12 @@ class BaseLocalSimulator(OpenQASMSimulator):
         self._validate_input_provided(circuit)
         BaseLocalSimulator._validate_shots_and_ir_results(shots, circuit.results, qubit_count)
 
-        operations = circuit.instructions
-        BaseLocalSimulator._validate_operation_qubits(operations)
-
         results = circuit.results
 
         simulation = self.initialize_simulation(
             qubit_count=qubit_count, shots=shots, batch_size=batch_size
         )
+        operations = circuit.instructions
         simulation.evolve(operations)
 
         if not shots:
@@ -493,12 +624,14 @@ class BaseLocalSimulator(OpenQASMSimulator):
         else:
             simulation.evolve(circuit.basis_rotation_instructions)
 
-        return self._create_results_obj(results, openqasm_ir, simulation, measured_qubits)
+        return self._create_results_obj(
+            results, openqasm_ir, simulation, measured_qubits, mapped_measured_qubits
+        )
 
     def run_jaqcd(
         self,
         circuit_ir: JaqcdProgram,
-        qubit_count: int,
+        qubit_count: Any = None,
         shots: int = 0,
         *,
         batch_size: int = 1,
@@ -508,7 +641,7 @@ class BaseLocalSimulator(OpenQASMSimulator):
         Args:
             circuit_ir (Program): ir representation of a braket circuit specifying the
                 instructions to execute.
-            qubit_count (int): The number of qubits to simulate.
+            qubit_count (Any): Unused parameter; in signature for backwards-compatibility
             shots (int): The number of times to run the circuit.
             batch_size (int): The size of the circuit partitions to contract,
                 if applying multiple gates at a time is desired; see `StateVectorSimulation`.
@@ -523,6 +656,10 @@ class BaseLocalSimulator(OpenQASMSimulator):
                 as a result type when shots=0. Or, if StateVector and Amplitude result types
                 are requested when shots>0.
         """
+        if qubit_count is not None:
+            warnings.warn(
+                f"qubit_count is deprecated for {type(self).__name__} and can be set to None"
+            )
         self._validate_ir_results_compatibility(
             circuit_ir.results,
             device_action_type=DeviceActionType.JAQCD,
@@ -531,6 +668,8 @@ class BaseLocalSimulator(OpenQASMSimulator):
             circuit_ir,
             device_action_type=DeviceActionType.JAQCD,
         )
+        qubit_map = BaseLocalSimulator._map_circuit_to_contiguous_qubits(circuit_ir)
+        qubit_count = len(qubit_map)
         BaseLocalSimulator._validate_shots_and_ir_results(shots, circuit_ir.results, qubit_count)
 
         operations = [
@@ -540,8 +679,6 @@ class BaseLocalSimulator(OpenQASMSimulator):
         if shots > 0 and circuit_ir.basis_rotation_instructions:
             for instruction in circuit_ir.basis_rotation_instructions:
                 operations.append(from_braket_instruction(instruction))
-
-        BaseLocalSimulator._validate_operation_qubits(operations)
 
         simulation = self.initialize_simulation(
             qubit_count=qubit_count, shots=shots, batch_size=batch_size