+++ test/unit_tests/braket/analog_hamiltonian_simulator/test_get_coefs.py
@@ -94,7 +94,7 @@ def test_get_func(para):
             + (values[ind + 1] - values[ind]) / (times[ind + 1] - times[ind]) * (t - times[ind])
         )
 
-    assert all([item_1 == item_2 for item_1, item_2 in zip(vals, trueval)])
+    assert np.allclose(vals, trueval)
 
 
 @pytest.mark.parametrize(